Moved statink_uploader.py from ikalog/outputs to ikalog/utils
----
This is one of PR requests to create a command line tool to upload
a stat.ink payload to the server.

After these changes, we will be able to do 2 pass process:
1. analysis of the video and save a stat.ink payload file.
2. upload the payload file anytime.

Here's the whole changes to be requested so far.
https://github.com/hiroyuki-komatsu/IkaLog/commits/statink
